Moves Point2D and Point3D to primitives directory
