+++ classes/Sprite.ts
@@ -1,5 +1,5 @@
 import Object2D from './Object2D';
-import { Point2D } from './Point2D';
+import { Point2D } from './primitives/Point2D';
 
 export class Sprite extends Object2D {
   public mesh: Point2D[];
